backend/pyproject.toml
[tool.black]
line-length = 80

[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Daniel Ribeiro <djbreves@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.11.2"
Authlib = "0.14.3"
fastapi = "0.101.0"
celery = "5.0.0"
redis = "3.5.3"
httpx = "0.24.1"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
requests = "2.24.0"
SQLAlchemy = "2.0.19"
passlib = "1.7.4"
sqlalchemy-utils = "0.41.1"
python-multipart = "0.0.5"
pyjwt = "1.7.1"
mangum = "^0.17.0"
aws-psycopg2 = "^1.3.8"
python-dotenv = "1.0.0"
aws-secretsmanager-caching = "^1.1.1.5"

[tool.poetry.group.dev.dependencies]
uvicorn = "0.23.2"
pytest = "^7.1.3"
ipython = "^7.31.1"
debugpy = "^1.6.7"
cfn-lint = "^0.79.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "app"
]